requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
jinja2>=3.1.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
itsdangerous>=2.2.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
pytest>=8.2.0
httpx>=0.27.0
pytest-asyncio>=0.23.0